+++ async_pokepy/__init__.py
@@ -10,10 +10,17 @@ A basic wrapper for the PokeAPI.co API.
 
 __title__ = "async_pokepy"
 __author__ = "Lorenzo"
+__docformat__ = "restructuredtext en"
 __license__ = "MIT"
 __copyright__ = "Copyright 2019 Lorenzo"
-__version__ = "0.0.9a"
+__version__ = "0.1.0a"
 
-from .client import Client
-from .exceptions import *
-from .types import *
+from collections import namedtuple
+
+from .client import Client  # noqa: F401
+from .exceptions import *  # noqa: F401
+from .types import *  # noqa: F401
+
+VersionInfo = namedtuple("VersionInfo", "major minor micro releaselevel")
+
+version_info = VersionInfo(major=0, minor=1, micro=0, releaselevel="alpha")