+++ example/example.py
@@ -8,7 +8,7 @@ async def main(query):
 
     pokemon = await client.get_pokemon(query)
 
-    fmt = " ".join(map(str, pokemon.abilities))
+    fmt = ", ".join(map(str, pokemon.abilities))
     print("{0} has the abilities {1}".format(pokemon, fmt))
 
     await client.close()
@@ -16,3 +16,5 @@ async def main(query):
 
 loop = asyncio.get_event_loop()
 loop.run_until_complete(main("Snorlax"))
+
+# This will output: "Snorlax has the abilities Gluttony, Thick Fat, immunity"